requirements.txt
ccxt
pandas
numpy
ta
requests
python-dotenv
